Move URL query parsing to HTTP edge
